[ISSUE-1254] Move JMAP settings APIs to jmap-extensions-api
